requirement.txt
asgiref==3.4.1
autopep8==1.5.7
Django==3.2.6
django-cors-headers==3.8.0
django-environ==0.4.5
django-filter==2.4.0
django-rest-passwordreset==1.2.0
djangorestframework==3.12.4
Pillow==8.3.1
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
